Created a Shader asset (a dumb GraphicsAPI handle wrapper). Finished implementing the ShaderTypeManager that performs loading and freeing of the Shader assets, when requested by the AssetManager.
Moved graphics API handles to a separate header file to avoid pulling in the entire GraphicsAPI header. This should allow me to speed up compilation in several places.

From now on, the createShader method takes a pointer to a memory buffer and its length instead of a file path. I'll do the same with images (textures) soon. File loading and management is the responsibility of the AssetManager. GraphicsAPI functions that interact with the filesystem were a mistake.