Moved the AssetHandle definition to a separate header. This allows me to avoid the inclusion of the template-heavy AssetManager header in some places.
Fixed a bug in the load method of the AssetManager and added a getSystemAsset() method that can be used to fetch a system asset by its filename.

Added an initMissingAssetHandle() method to the TypeManagerBase. TypeManagers used to fill the missingAssetHandles in their constructors, however, that introduced special cases and forced me to fill their metadata by hand (because it wasn't loaded by that point). From now on, things go like this: the AssetManager is initialized -> type managers are constructed -> the AssetManager loads the Metadata to the manifest (to be implemented) -> initMissingAssetHandle() is called.

From now on, all existing classes that require shaders request them from the AssetManager. This change introduced some unexpected execution order dependencies and I was forced to start initializing the AssetManager earlier in the Engine's init() method.

Added a release() method to the ReferenceCountedHandle. This allows me to avoid some potential problems with asset cleanup.

Some more minor changes.