Changed AssetManager's load() and getSystemAsset() methods to (optionally) return metadata. Added a method to obtain metadata for any asset that's registered in the manifest and a couple more helpers.
Moved the Mesh loader to assets/loader.

Changed the constructor of the CubemapSkybox. From now on, it takes a nameHash instead of a file path and uses an AssetHandle instead of a raw GraphicsAPI Image.

Started refactoring the GraphicsAPI Image creation functions. Just like with shaders, I want to make sure the classes that derive from the GraphicsAPI don't need to deal with the filesystem.

Created a Texture Asset and a TextureTypeManager.

Removed a few comments that (for some reason) were in my native Lithuanian.