Finised implementing my custom TextureLoader and started using it.
Removed my dependency on gli because it is no longer required.

Temporarily disabled the check for VK_KHR_DEDICATED_ALLOCATION_EXTENSION_NAME. It appears that it depends on yet another extension that I don't check for and Vulkan's validation layers complain about it. I'll need to address their warnings and re-enable it.

Created a buffer that I plan to use for batched image transfers.