Finally have shader macros and includes working and used in shader compilation.
Vertex data layout definitions and a few more constant arrays have been turned into static variables inside functions. This simplified the code in a few places and fixed a few more undefined initialization order bugs.

Shader generation and compilation results are now convertible to bool. This simplifies a code in a few places.

Updated the documentation in a few places.