From now on, derived nodes will have to add connectors using addInput() and addOutput() functions. They simplify the process and perform extra validation.
Since the vectors that store connectors are now private, I added a couple functions that can be used to retrieve references to said connectors. They're required when changing the mode has to change the type of the connectors as well.

Added a ton of new MaterialNode types. Moreover, they now all derive from a single non templated class MaterialNodeBase that I may (or may not) use for some common serialization functions later. I've also created a helper class GenTypeMaterialNode that signifficantly simplifies the creation of nodes with modes that change the types of inputs and outputs.

Added localization strings required by the new MaterialNode types.