Tons of material node editor improvements:
Nodes are now grouped based on type or function. Added methods that can be used to create and retrieve said groups and a ton of validation.

LogicGraphNodeTypeInfo now includes the node type. This simplifies the code in a few places.

Added checks for disabled node input and output connectors.

Disabled connectors are now displayed in a different colour.

Reduced the default width of the nodes. Minimal allowed scale is once again back to 0.5.

It's now possible to see the documentation or change the type of the node by right clicking on it.

Almost all GLSL functions I care about are now available as nodes and there's a ton of them. Actual code generation is going to be a pain.

Fixed some minor bugs, refactored some frequently used code into separate functions and removed some dead code.