The TextSerializable interface now has a helper function called getJSONString(). It automatically serializes an object that implements TextSerializable into a pretty printed JSON string.
LogicGraphNode and LogicGraph now implement TextSerializable and may be serialized to JSON. I've also added isDeserializing parameters to a couple functions in these classes because deserialization requires slightly different logic.

The users of the LogicGraph now need to implement addNodeImpl() instead of addNode() because, thanks to changes in logic caused by deserialization, the node key needs to be determined in advance. That's what addNode() does before calling addNodeImpl()

Fixed a node name editing bug (the logic was relying on a variable that I was no longer setting and using).