Implemented node graph conversion to GLSL code.
Added a couple more shader macros that are used in shader generation.

Moved renderer properties to separate classes that can be accessed using a static function. This will allow me to implement renderer independent shader generation and compilation.

Fixed a couple bugs and added a ton of extra localization strings for the new node editor nodes.