Added a != operator to all metadata classes.
Implemented metadata serialization and deserialization tests and used them to fix a bug in the MetadataBase class serialization.

Moved certain logging implementation details from the header into the cpp file, which signifficantly reduced the number of headers that get included with the logging header. This once again speeds up compilation in multiple places.